Show labels unchecked when "Open" button selected
The instance of LabelsMenu used with the "Open" button should show all
available labels in an unselected state. The user can then check labels
to filter the list of available models to open.

Note that the selection state of the labels for the current model isn't
modified; the labels are only displayed unchecked in the LabelsMenu
widget.