Break LabelsMenu into classes for open, save and manage
LabelsMenu was getting unwieldy with many conditionals for the open,
save and manage actions. LabelsMenu is now used for manage. The open
and save actions have their own LabelsMenu classes.

Adding/deleting labels has been removed from open and save menus.

This also fixes a bug with the open droplist not repopulating when
labels are unselected.