Add handlers for question links to sign-in screen
The "New User?" and "Forgot Password?" questions on the sign-in screen
now have ClickHandlers that display information.

Also changed vertical alignment of content: instead of center-aligned,
it's now top-aligned with an offset calculated from the browser window
height.

The width of the content is also constrained by the browser window
width.