Disable window close event in sign-in screen
We've been trapping the window close event and reminding users that
info will be lost if the leave the page. I've changed this so the
handler is now associated with the Perspective; it's attached when the
user signs in, and removed when the user signs out. It won't trap a
refresh/close of the sign-in page.