Only show global parameters on the driver component
For example, `run_duration` is shown only in the driver component.
All the passenger components use this value (because we want the run
duration to be the same for all components), but it's not shown.

I had to change ParameterTable#loadTable to take as input the current
ComponentCell instead of just the current componentId. This links the
ParameterTable to the ModelTree, which may not be a bad thing. I updated
references to loadTable in other classes.