On file upload, prompt user to save model only if model hasn't been saved previously
Previously, I had prompted the user to save before every file
upload. This isn't necessary, though; the model only needs to be saved
to the server, such that it has been assigned an id and it has a folder
in the server's **uploads** directory. Also, as @iovereem points out,
it's annoying to get prompted for every save.